feat: ensure the hosts use `network-scripts`
Rocky Linux 8 comes preconfigured with `NetworkManager` for handling the
network capabilities of the operating system. We however, would prefer
to use `network-scripts` instead, therefore, we must ensure that the
package is installed and the `network` service is selected over
`NetworkManager`.

NOTE: `network-scripts` is deprecated and has been removed from Rocky
Linux 9.

NOTE: For operating systems such as Ubuntu we prefer to use
`systemd-networkd` support for which will be added at a later date.